Vec2 time !!!
Changed most IVec2 to Vec2

+++ src/main/utils/noise/PseudoRandom.java
@@ -8,7 +8,7 @@ public class PseudoRandom {
         double randFactor= 0.5+(noiseGenerator.noise(a,b,0)/2);
         return Math.clamp(Math.round(randFactor*(maxX-minX))+minX,minX,maxX);
     }
-    public static int getRandomBetween(int minX, int maxX, int a, int b, int size){
+    public static int getRandomBetween(int minX, int maxX, double a, double b, int size){
         double randFactor= 0.5+(noiseGenerator.noise(a,b,0,size)/2);
         return Math.clamp(Math.round(randFactor*(maxX-minX))+minX,minX,maxX);
     }
